LS980: USB: HSIC SMSC HUB: Pass AUXDATA while adding HSIC_HOST
AUXDATA should be passed to update HSIC device name. This
is required for getting clock resources and userspace apps
that uses device name for proper functioning. In case of
HSIC connection using SMSC HUB, HSIC host is added as a
child node of SMSC HUB. Hence, pass auxdata for SMSC_HUB
which points to the child node's (HSIC_HOST) auxdata.

Change-Id: I8b763341a845fbfba9fc087bbc59bfc677bc3c04
Signed-off-by: Manu Gautam <mgautam@codeaurora.org>